#6: removed windows types from interfaces, moved thread/events/lock into support classes

+++ src/DataDistributionManagerKafka/dllmain.cpp
@@ -20,7 +20,7 @@
 #include "stdafx.h"
 
 BOOL APIENTRY DllMain( HMODULE hModule,
-                       DWORD  ul_reason_for_call,
+                       unsigned long  ul_reason_for_call,
                        LPVOID lpReserved
 					 )
 {
